requirements.txt
langchain-openai
langchain_core
python-dotenv
langchain_google_firestore 
ragstack-ai